Brutto finish of header+nav, hero and quote sctn
At this point in time, the header with logo and navigation is in place, but navigation is not responsive for small devices. Hero section is finished as far as my current designing skills can get me. Quote section was easy, but I would like to add some quote image icons later, but now I need to focus on more progress on the whole page or I will mentally break down haha